Create bitflip proof of concept
This commit is a proof of concept of a situation where the payload size
in the header of a message is corrupted.  Despite that corruption, the
peer is not disconnected, nor are future messages misaligned; the peer
can continue working as normal.

This is done by sending two carefully crafted but otherwise legitimate
bitcoin messages one after another.  Recall that we don't disconnect on
an invalid command or an invalid checksum.  The messages have been
crafted in the following way:

Message 1:
 - Last 4 bytes of the payload are the same as the netmagic
 - Payload size field originally has a 1 in the 6th bit
 - Through sending, that bit is flipped

Message 2:
 - Has no payload
 - Command name is 8 or fewer bytes
 - Through sending, the 6th bit in the 9th byte of the command name is
   flipped

Here are hexdumps of the messages before and after the bitflips

Original Message 1

Offset(h) 00 01 02 03 04 05 06 07

00000000  FA BF B5 DA 73 6F 6D 65  ú¿µÚsome
00000008  64 61 74 61 00 00 00 00  data....
00000010  0C 00 00 00 81 8F D3 B6  ......Ó¶
00000018  73 6F 6D 65 6A 75 6E 6B  somejunk
00000020  FA BF B5 DA              ú¿µÚ

Modified Message 1

Offset(h) 00 01 02 03 04 05 06 07

00000000  FA BF B5 DA 73 6F 6D 65  ú¿µÚsome
00000008  64 61 74 61 00 00 00 00  data....
00000010  08 00 00 00 81 8F D3 B6  ......Ó¶
00000018  73 6F 6D 65 6A 75 6E 6B  somejunk
00000020  FA BF B5 DA              ú¿µÚ

Original Message 2

Offset(h) 00 01 02 03 04 05 06 07

00000000  FA BF B5 DA 65 6D 70 74  ú¿µÚempt
00000008  79 6D 73 67 00 00 00 00  ymsg....
00000010  00 00 00 00 5D F6 E0 E2  ....]öàâ

Modified Message 2

Offset(h) 00 01 02 03 04 05 06 07

00000000  FA BF B5 DA 65 6D 70 74  ú¿µÚempt
00000008  79 6D 73 67 04 00 00 00  ymsg....
00000010  00 00 00 00 5D F6 E0 E2  ....]öàâ